+++ content/web/Landing Pages/This is a test folder/update-seo-info copy.md
@@ -1,0 +1,34 @@
+# Updating Landing Page SEO Information
+
+## Intro
+Updating SEO information for Landing Pages is slightly different than for other pages on liferay.com.
+
+Landing pages have URLs like these:
+- <code>www.liferay.com/l/my-landing-page</code>
+- <code>www.liferay.com/resources/l/my-landing-page</code>
+
+## Edit SEO Info
+
+1. Navigate to the landing page you want to update.
+
+2. Scroll to the bottom of the page.
+
+3. <img src="/images/web/Landing_Pages/landing-page-seo-edit-landing-page.png">
+	Click "Edit Landing Page".
+</img>
+
+4. <img src="/images/web/Landing_Pages/landing-page-seo-click-custom-fields.png">
+	Choose "Custom Fields" from the side menu.
+</img>
+
+5. <img src="/images/web/Landing_Pages/landing-page-seo-select-translation.png">
+	Choose the flags corresponding to the translation you are updating.
+</img>
+
+6. <img src="/images/web/Landing_Pages/landing-page-seo-update-custom-fields.png">
+	Update the SEO information.
+</img>
+
+7. <img src="/images/web/Landing_Pages/landing-page-seo-publish.png">
+	Publish your changes.
+</img>
